@@ -16,8 +16,7 @@ Gem::Specification.new do |s|
   s.files = Dir["{app,config,db,lib}/**/*"] + ["MIT-LICENSE", "Rakefile", "README.md"]
   s.test_files = Dir["test/**/*"]
 
-  s.add_dependency "rails", "~> 4.0"
-  # s.add_dependency "jquery-rails"
+  s.add_dependency "rails", "~> 5.0"
 
   s.add_development_dependency "sqlite3"
   s.add_dependency "ruby-openid", "~> 2.0"
